Drop Rust 1.25 support for pointer addition
Previously I had wanted to support older style code syntax required in Rust
1.25.  But with the addition of `CodepointIterator` I needed pointer
addition support and that wasn't stablized until 1.26.  Adding a flag
for unstable features would only work with a nightly branch which would
be a problem for programs using this library.  It's simpler to just move
to supporting the current stable Rust.