Keep `Value` as a private API
Clean up any documentation or use of `Value` from the public API.

`VM::protect` is going to be recommended more and more so I've changed
it from taking a function that returns a Value to once that returns an
AnyObject.  This is a breaking change in the public API.

Removed any use of the `.value()` method in the public API
documentation.

`NilClass` has had `Copy` and `Clone` derived on it.

Added readme section for Ruby's Future and SemVer.