Change `Object#send` to `unsafe`
Because exception raising in Ruby can crash the Rust code if it isn't
handled via protect then we should consider `Object#send` as unsafe
because invalid arguments can be given to raise an exception.

When wraping code with `unsafe { ... }` the developer is stating that
they've taken the appropriate measures to make the context outside of
the unsafe block safe by what was done inside it.  In this projects case
though it's okay to put the unsafe block inside the `VM::protect`
closure as it achieves the same result.