Use native Ruby function for getting a string's length
Without this change, Ruby 3.2 was failing to correctly provide me with strings, because the length was always set to zero.

I'm not sure exactly what the change in 3.2 that triggered the problem, but given that rutie's `rstring_len` appears to be an implementation of a function in CRuby that's named `rbimpl_` (which is a namespace meaning "no compatibility should be expected", I don't think this is Ruby's fault.